Several improvements to the slide.js viewer
The vcentered class now uses a CSS transform trick instead of JS to vertically
center content. This works better for me than the JavaScript, which was quirky.

Fixed a few font display issues in the CSS.

If IDs are set on the current slide, the URL hash will be set to that ID and the
slides will start at that position if a direct link is used, which is super
handy for testing while editing.